+++ src/reportlab/rl_settings.py
@@ -159,9 +159,11 @@ reserveTTFNotdef=1                                  #if true force subset elemen
                                                     #PDFUA forbids index 0(.notdef) in strings
 documentLang=None                                   #pdf document catalog Lang value xx-xx not ee_xx
 encryptionStrength=40                               #the bits for standard encryption 40, 128 or 256 (AES)
-trustedHosts=None                                   #set to a list of trusted for access hosts None means
-                                                    #all are trusted glob patterns eg *.reportlab.com are
+trustedHosts=None                                   #set to a list of trusted for access hosts
+                                                    #glob patterns eg *.reportlab.com are
                                                     #allowed. In environment use a comma separated string.
+                                                    #to use data: or file: schemes trustedHosts must contain localhost
+                                                    #None or other false value means no hosts are trusted
 trustedSchemes=['file', 'rml', 'data', 'https',     #these url schemes are trusted
                 'http', 'ftp']
 renderPMBackend='rlPyCairo'                         #or '_renderPM' if available